requirements.txt
osmnx
gym
pytorch
pytorch-geometric
networkx
geopandas